Add open source fonts
Gill Sans and Palatino are not open source fonts and can not be relied upon
for building on Travis. This commit switchs the main beamer font to
TeXGyrePagella and the default sans serif font to OpenSans.

I am not entirely sure how to set these up as I could only get it to work
by adding \defaultfontfeatures to the style file to declare them so that
xelatex would find them in the TEXMF tree.