Redefine page layout and associated headers
This commit splits the header definition into two parts.
The left header is the logo and horizontal rule and the
right header is now the document description text.
Previously they were a single left header but this led to
spacing anomalies between pages where the spacing between
the top of the logo and the top of the page could vary and the
spacing between the ruler and the document identifier.
Separating these into two unrelated entities seems to have
stabilized the spacing issues.

In order to simplify the investigation, the page layout
has now been defined using the geometry package rather
than distinct settings that may interact unexpectedly.
This has the side effect that the margins can now
easily be displayed for debugging purposes using the
"showframe" option.

All page layout was tweaked to make the spacing correct
for the header. This led to the footer spacing now being
more consitent, although page numbers do seem to move
around from page to page.