Dynamically calculate lengths in header rules
Now that the left header is defined as a minipage, rather than
use fixed lengths of rulers to try to fudge the position of the
LSST text under the logo, calculate the values from the specified
width of the new minipage and the width of the logo. This ensures
that the LSST text is more or less centred on the logo. If the
logo is changed the text will move accordingly. The rulers will
also adjust.