Directly place logo title page
Remove the fragile \putlogo command and replace it with
direct placement of the logo and LSST text on the title page.
This has the advantage of allowing the logo to be positioned
without adjustments shifting the rest of the title content,
and removes any complaints about overfull hboxes because the
full paper width text box is now not part of the page itself.