Create Dockerfile for lsstsqre/lsst-texmf image
This Dockerfile effectively containerizes the texmf directory. This
allows a user or CI environment to use lsst-texmf as a self-contained
appliance for building a LSST LaTeX document.

The Dockerfile copies in only the texmf directory and sets the TEXMFHOME
environment variable.

This Dockerfile uses https://github.com/lsst-sqre/lsst-texlive as a
base; that image provides a containerized texlive distribution.