Move all build steps to Travis 'script' phase
This will ensure that the build is marked as a failure if there is an
issue with the  Docker build/upload.

Before only issues with the core latex builds (e.g. `make`) could have
triggered a build failure that would block a merge.

Note I'm keeping LSST the Docs documentation site upload in the
after_success phase because of a timeout issue (LTD Keeper needs to
become async).