Split scan into known acronyms and possible acronyms
First do a regex of all the acronyms we know of in the list that
do not look like standard acronyms. Then look at text to analyze
and extract all items that look like all upper case acronyms
of 2 or more characters.

Report acronyms that we find but which do not have definitions.