Prune Gaia acronym list.
There were many acronyms here which obviously don't apply to LSST, and often
match in misleading ways with acronyms which do apply to LSST (my favourite:
“LSP: Low System Priority”).

This is the result of a quick pass through to remove entries which probably
aren't useful in the LSST context. Given the length of the acronym list, I
have preferred speed over accuracy: it may be that some useful information is
lost, but I think it's preferable to re-add that when we need it rather than
stick with our current multitude of false matches.